Extract `theme` defition to `lib/theme.ts`
